move config functions from util.py to config.py
