+++ mock/tests/test_package_manager.py
@@ -7,7 +7,7 @@ from unittest import mock
 from unittest.mock import MagicMock
 
 from mockbuild.text import TemplatedDictionary
-from mockbuild.util import load_defaults
+from mockbuild.config import load_defaults
 from mockbuild.buildroot import Buildroot
 from mockbuild.package_manager import _PackageManager, Dnf
